requirements.txt
# Web framework + UI
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.36.0
python-multipart==0.0.9
python-dotenv==1.0.1

# PDF processing
PyMuPDF==1.24.10
pdfplumber==0.11.4

# Vector store
faiss-cpu==1.8.0
numpy==1.26.4

# HTTP client (Ollama + optional BLIP API)
requests==2.32.3

# Image captioning (optional — only needed for caption_mode=sample/full)
huggingface-hub==0.24.5
Pillow==10.4.0

# Ollama runs as a local service — install separately from https://ollama.com
# Then: ollama pull nomic-embed-text && ollama pull gemma3